backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djlint==1.36.4
gunicorn==25.3.0
prometheus_client==0.25.0
psycopg2-binary==2.9.12
python-decouple==3.8
django-redis==6.0.0
drf-spectacular==0.29.0
argon2-cffi==25.1.0
django-ipware==7.0.1
user-agents==2.2.0
geoip2==5.2.0